+++ lib/browser/base.rb
@@ -84,6 +84,12 @@ module Browser
         detect_version?(full_version, expected_version)
     end
 
+    # Detect if browser is Snapchat.
+    def snapchat?(expected_version = nil)
+      Snapchat.new(ua).match? &&
+        detect_version?(full_version, expected_version)
+    end
+
     # Detect if browser if Facebook.
     def facebook?(expected_version = nil)
       Facebook.new(ua).match? &&
@@ -99,7 +105,7 @@ module Browser
     # Detect if browser is WebKit-based.
     def webkit?(expected_version = nil)
       ua =~ /AppleWebKit/i &&
-        !edge? &&
+        (!edge? || Edge.new(ua).chrome_based?) &&
         detect_version?(webkit_full_version, expected_version)
     end